+++ middlewares/ratelimit.js
@@ -1,5 +1,5 @@
 const rateLimit = require("express-rate-limit");
-const {Server} = require('../models/server.model');
+const {Server} = require('../models/ServerSchema');
 
 // non-premium servers can handle ~5 verification per 30 minutes
 // premium servers can handle over 20 verification per 30 minutes
