+++ routes/server.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const serverController = require('../controllers/server.controller');
+const serverController = require('../controllers/server');
 const router = express.Router()
 
 router.get('/', serverController.getAll)
